If no volume, don't delete volume in teardown
If exception is thrown in setupClass before volume is created,
tearDownClass needn't delete volume.

Closes-Bug: 1350167

+++ tempest/api/compute/v3/servers/test_server_rescue_negative.py
@@ -56,7 +56,8 @@ class ServerRescueNegativeV3Test(base.BaseV3ComputeTest):
 
     @classmethod
     def tearDownClass(cls):
-        cls.delete_volume(cls.volume['id'])
+        if hasattr(cls, 'volume'):
+            cls.delete_volume(cls.volume['id'])
         super(ServerRescueNegativeV3Test, cls).tearDownClass()
 
     def _detach(self, server_id, volume_id):
